refactor: 구매 내역에 관련된 state와 함수를 커스텀 훅 형태로 분리
